big refactory. bug fixes. NPE at loading trainers.

+++ src/main/java/org/ArgentumOnline/server/ObjectInfo.java
@@ -104,7 +104,7 @@ public class ObjectInfo implements Constants {
     public short LingO;
     
     public short LingP;
-    public short Madera;
+    public int Madera;
     
     public short SkHerreria;
     public short SkCarpinteria;
@@ -333,7 +333,7 @@ public class ObjectInfo implements Constants {
 
         this.SkCarpinteria = ini.getShort(section, "SkCarpinteria");
         if (this.SkCarpinteria > 0) {
-            this.Madera  = ini.getShort(section, "Madera");
+            this.Madera  = ini.getInt(section, "Madera");
         }
 
         if (this.ObjType == OBJTYPE_BARCOS) {
